% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lrt}
\alias{lrt}
\title{Teste da razão de verossimilhança generalizada}
\usage{
lrt(f, data, kicks, par0 = NULL, ...)
}
\arguments{
\item{f}{Função densidade de probabilidade considerada no teste da razão de verossimilhança. Essa função deverá ser
implementada conforme os exemplos.}

\item{data}{Um conjunto de dados que será considerado para a realização do teste.}

\item{kicks}{Chutes iniciais que desejamos considerar no método de otimização.}

\item{par0}{Uma lista contendo como primeiro elemento um vetor com o nomes das variáveis que serão fixadas como hipótese nula
(variáveis aos quais desejamos testar) e um vetor com os valores fixados para cada uma das respectivas variáveis.}

\item{...}{Lista de argumentos adicionais que serão passados à função \code{optim()} otimizada para otimização. Por exemplo,
será possível escolher o método de otimização a ser utilizado.}

\item{par1}{Lista com dois elementos, sendo o primeiro um vetor com os nomes das variáveis que receberão valores fixos sob a
hipótese alternativa e o segundo elemendo é um outro vetor com os valores impostos às variáveis.}
}
\value{
Valor da estatística do teste da razão de verossimilhança generalizada.
}
\description{
Calcula a estatística da razão de verossimilhança generalizada.
}
\details{
A função recebe como argumento uma função densidade de probabilidade ou uma função de probabilidade
que é passada como argumento à \code{f}. O objeto passado como argumento de \code{f} deverá ser implementado conforme os
exemplos. Note que se a função for implementada segundo os exemplos, não haverá a necessidade de implementar a função sob a
hipótese nula, visto que são as mesmas funções. Para especificar os parâmetros que serão fixados, i.e, para especificar
a distribuição sob a hipótese nula, utiliza-se o argumento \code{par0} que receberá uma lista formada por dois vetores.
O primeiro vetor da lista de verá ser um vetor de strings com os nomes das variáveis que deseja-se fixar e o segundo vetor
deverá conter os valores que serão atribuidos à cada uma das variáveis passada ao primeiro vetor.
}
\examples{
 pdf_w <- function(par, x, var = NULL){
     alpha <- par[1]
     beta <- par[2]
     dweibull(x, shape = alpha, scale = beta)
  }

  rw <- function(n = 1L, alpha, beta){
   rweibull(n = n, shape = alpha, scale = beta)
  }

  data <- rw(n = 100L, alpha = 1, beta = 1)

  lrt(f = pdf_w, data = data, kicks = c(1, 1), par0 = list("beta", 1))
}
\author{
Pedro Rafael D. Marinho
}
